fix featrue/auth frontend file
